# docs/swagger.yaml
openapi: 3.0.3
info:
    title: ServeBin
    description:
        Welcome to ServeBin documentation! ServeBin is a cutting-edge HTTP
        testing and debugging tool, built with the latest technologies in Go. This documentation
        provides comprehensive details about the endpoints, parameters, and responses
        offered by ServeBin, empowering developers to streamline their testing workflows
        and ensure the reliability of their applications. Explore the various features
        and capabilities of ServeBin to optimize your development process and elevate
        your HTTP testing experience.
    termsOfService: https://servebin.dev/
    contact:
        email: contact@mrayush.me
    license:
        name: BSD 3-Clause
        url: https://github.com/AyushAgnihotri2025/ServeBin/blob/master/LICENSE
    version: v1.3.0
servers:
    - url: https://servebin.dev/
    - url: https://s1.servebin.dev/
    - url: https://s2.servebin.dev/
...
paths:
...
    /get:
        get:
            tags:
                - HTTP Methods
            summary: Returns the request parameters.
            description:
                Returns different type of request parameters like form data, json
                data, raw data, headers etc.
            parameters:
                - name: customheader
                  in: header
                  description: Header
                  schema:
                      type: string
                  example: servebin-demo
                - name: queryparam
                  in: query
                  description: Query Paramater
                  schema:
                      type: string
                  example: servebin
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.EmptyResponse"
                "400":
                    description: Bad Request
                    content: {}
                "404":
                    description: Not Found
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
    /patch:
        patch:
            tags:
                - HTTP Methods
            summary: Returns the request parameters.
            description:
                Returns different type of request parameters like form data, json
                data, raw data, headers etc.
            parameters:
                - name: customheader
                  in: header
                  description: Header
                  schema:
                      type: string
                  example: servebin-demo
                - name: queryparam
                  in: query
                  description: Query Paramater
                  schema:
                      type: string
                  example: servebin
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                body:
                                    type: string
                                    description: Body
                                formdata:
                                    type: string
                                    description: Form Data
                                    format: binary
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.BodyDataResponse"
                "400":
                    description: Bad Request
                    content: {}
                "404":
                    description: Not Found
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /post:
        post:
            tags:
                - HTTP Methods
            summary: Returns the request parameters.
            description:
                Returns different type of request parameters like form data, json
                data, raw data, headers etc.
            parameters:
                - name: customheader
                  in: header
                  description: Header
                  schema:
                      type: string
                  example: servebin-demo
                - name: queryparam
                  in: query
                  description: Query Paramater
                  schema:
                      type: string
                  example: servebin
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                body:
                                    type: string
                                    description: Body
                                formdata:
                                    type: string
                                    description: Form Data
                                    format: binary
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.BodyDataResponse"
                "400":
                    description: Bad Request
                    content: {}
                "404":
                    description: Not Found
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
    /put:
        put:
            tags:
                - HTTP Methods
            summary: Returns the request parameters.
            description:
                Returns different type of request parameters like form data, json
                data, raw data, headers etc.
            parameters:
                - name: customheader
                  in: header
                  description: Header
                  schema:
                      type: string
                  example: servebin-demo
                - name: queryparam
                  in: query
                  description: Query Paramater
                  schema:
                      type: string
                  example: servebin
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                body:
                                    type: string
                                    description: Body
                                formdata:
                                    type: string
                                    description: Form Data
                                    format: binary
            responses:
                "200":
                    description: OK
                    content:
                        "*/*":
                            schema:
                                $ref: "#/components/schemas/response.BodyDataResponse"
                "400":
                    description: Bad Request
                    content: {}
                "404":
                    description: Not Found
                    content: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/response.HTTPError"
...
components:
    schemas:
        response.BodyDataResponse:
            type: object
            properties:
                arg:
                    type: object
                    additionalProperties: true
                data:
                    type: object
                files:
                    type: object
                form:
                    type: object
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
                json:
                    type: object
                method:
                    type: string
                url:
                    type: string
...
        response.EmptyResponse:
            type: object
            properties:
                arg:
                    type: object
                    additionalProperties: true
                headers:
                    type: object
                    additionalProperties:
                        type: string
                ip:
                    type: array
                    items:
                        type: object
                method:
                    type: string
                url:
                    type: string
...
        response.HTTPError:
            type: object
            properties:
                code:
                    type: integer
                    example: 400
                message:
                    type: string
                    example: status bad request
...